# specs/001-merchant-review-insights/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Whole Appliance Review Insights API
  version: 2.1.0
servers:
  - url: http://localhost:8000/api/v1
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        '200':
          description: Service status and active Gold run
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Health'}
  /products:
    get:
      operationId: listProducts
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: category, in: query, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
        - {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
      responses:
        '200':
          description: Only fully processed products from the active Gold run
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProductPage'}
  /products/{parentAsin}:
    get:
      operationId: getProduct
      parameters:
        - {$ref: '#/components/parameters/ParentAsin'}
      responses:
        '200':
          description: Product overview
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Product'}
  /products/{parentAsin}/themes:
    get:
      operationId: listThemes
      parameters:
        - {$ref: '#/components/parameters/ParentAsin'}
        - {name: sentiment, in: query, schema: {enum: [positive, negative]}}
      responses:
        '200':
          description: Positive and negative evaluation themes derived only from ABSA
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Theme'}
  /products/{parentAsin}/analysis-status:
    get:
      operationId: getProductAnalysisStatus
      parameters:
        - {$ref: '#/components/parameters/ParentAsin'}
      responses:
        '200':
          description: Completion and empty-result status for all three analysis facets
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ProductAnalysisFacet'}
  /themes/{themeId}/reviews:
    get:
      operationId: listThemeReviews
      parameters:
        - {name: themeId, in: path, required: true, schema: {type: string}}
        - {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
      responses:
        '200':
          description: Original review evidence for one theme
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/ReviewEvidence'}
  /products/{parentAsin}/improvements:
    get:
      operationId: listImprovements
      parameters:
        - {$ref: '#/components/parameters/ParentAsin'}
        - {name: source, in: query, schema: {enum: [explicit_suggestion, implicit_problem]}}
      responses:
        '200':
          description: Improvement themes derived from explicit suggestions and implicit problems
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Theme'}
components:
  parameters:
    ParentAsin:
      name: parentAsin
      in: path
      required: true
      schema: {type: string}
  schemas:
    Health:
      type: object
      required: [service, active_run]
      properties:
        service: {enum: [ok, degraded]}
        active_run: {type: [string, 'null']}
    ProductPage:
      type: object
      required: [items, page, page_size, total]
      properties:
        items: {type: array, items: {$ref: '#/components/schemas/Product'}}
        page: {type: integer}
        page_size: {type: integer}
        total: {type: integer}
    Product:
      type: object
      required: [parent_asin, title, category, review_count, analysis_status]
      properties:
        parent_asin: {type: string}
        title: {type: string}
        category: {type: string}
        review_count: {type: integer}
        average_rating: {type: [number, 'null']}
        analysis_status: {enum: [pending, processing, ready, failed]}
    Theme:
      type: object
      required: [theme_id, theme_type, name, sentiment, sources, review_count, share, trend]
      properties:
        theme_id: {type: string}
        theme_type: {enum: [evaluation, improvement]}
        name: {type: string}
        sentiment:
          type: [string, 'null']
          enum: [positive, negative, null]
          description: Required for evaluation themes and null for improvement themes
        sources:
          type: array
          items: {enum: [absa_positive, absa_negative, explicit_suggestion, implicit_problem]}
        review_count: {type: integer}
        share: {type: number}
        naming_method: {enum: [template, center_phrase]}
        trend:
          type: array
          items: {$ref: '#/components/schemas/TrendPoint'}
    ProductAnalysisFacet:
      type: object
      required: [parent_asin, theme_type, sentiment, status, theme_count, reason]
      properties:
        parent_asin: {type: string}
        theme_type: {enum: [evaluation, improvement]}
        sentiment:
          type: [string, 'null']
          enum: [positive, negative, null]
        status: {enum: [pending, processing, ready, failed]}
        theme_count: {type: integer, minimum: 0}
        reason:
          type: [string, 'null']
          enum: [no_qualified_theme, null]
    ReviewEvidence:
      type: object
      required: [review_id, text, rating, review_month]
      properties:
        review_id: {type: string}
        text: {type: string}
        evidence_sentence: {type: [string, 'null']}
        rating: {type: number}
        review_month: {type: [string, 'null']}
    TrendPoint:
      type: object
      required: [month, review_count, share]
      properties:
        month: {type: string, pattern: '^\\d{4}-(0[1-9]|1[0-2])$'}
        review_count: {type: integer}
        share: {type: number}